+++ packages/experimental/Menu/src/context/menuContext.ts
@@ -13,13 +13,14 @@ export interface MenuContextValue extends MenuState {
 
 export const MenuContext = React.createContext<MenuContextValue>({
   isControlled: false,
-  checked: {},
-  defaultChecked: {},
+  checked: [],
+  defaultChecked: [],
   hasCheckmarks: false,
   isSubmenu: false,
   open: false,
   onCheckedChange: () => false,
   setOpen: () => false,
+  shouldFocusOnContainer: false,
   triggerRef: null,
 });