+++ packages/components/Checkbox/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/checkbox
 
-This log was last generated on Tue, 07 Jun 2022 22:59:24 GMT and should not be manually modified.
+This log was last generated on Mon, 13 Jun 2022 23:58:18 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.15.12
+
+Mon, 13 Jun 2022 23:58:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.16.3
+- Bump @fluentui-react-native/pressable to v0.8.45
 
 ## 0.15.11